+++ ReleaseNotes.md
@@ -11,6 +11,7 @@
   VDAT (issue #1963)
 * added --rhozp option to astrometry rotuines, to allow user to specify
   the rotation angle of the rho stage 
+* Add pyhetdex.tools.db_helpers (from VDAT, issue #2026)
 
 ## Development version @ branches/selection_function_devel